requirements.txt
pycrypto==2.6.1
pycryptodomex==3.18.0
python-dotenv==1.0.0
sockets==1.0.0

